Just a tiny piece of code refactoring
